Fix Python 3 unittest heisenbug, but...
This fixes an ordering issue on keyslice_pairs when using replace= by
sorting the input. But it does not really fix the underlying problem,
which is that 1) keyslice_pairs needs more rigor in its input definition:
are these unique pairs? and 2) replacements need to be handled completely
before yielding other pairs.