only give a list to _set_flat()
when giving a generator (or iterator) to _set_flat(),
it might malfunction as it might try to fully iterate it more than once.

with python 3, a generator is returned from pairs.items() [directly
above the fixed line], thus set_flat() might malfunction when a dict
is given to it (instead of a list of pairs).